requirements.txt
Django==5.1.1
Pillow==10.4.0
crispy-tailwind==1.0.3
djangorestframework==3.15.2
django-filter==24.3
Markdown==3.7
django-environ==0.11.2
psycopg2-binary==2.9.10